# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yoko API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users (Route only for admins)",
                "operationId": "a022ddae96a9d9fa2b5fa09801eb6c34",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query for searching particular user by email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Request successful"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "currentPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "lastPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "size": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "users": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "MaksymProkopenko@gmail.com"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Maksym"
                                                            },
                                                            "surname": {
                                                                "type": "string",
                                                                "example": "Prokopenko"
                                                            },
                                                            "isBlocked": {
                                                                "type": "boolean",
                                                                "example": "0"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/users/collections": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all user collections",
                "operationId": "f40e2bf41cb3f637d8ea8c9f6cf7c7e2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query for searching collections by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Request successful"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "currentPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "lastPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "size": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Tourism"
                                                            },
                                                            "text": {
                                                                "type": "string",
                                                                "example": "Discover the enchanting beauty of Italy's Amalfi Coast. With its picturesque cliffside villages, azure waters, and vibrant culture, the Amalfi Coast offers an unforgettable travel experience. Whether you're exploring the historic streets of Positano, indulging in delicious Mediterranean cuisine, or lounging on pristine beaches, every moment is a postcard-worthy memory waiting to be captured."
                                                            },
                                                            "translationUk": {
                                                                "type": "string",
                                                                "example": "Відкрийте для себе чарівну красу узбережжя Амальфі в Італії. З його мальовничими селами на скелях, блакитними водами та яскравою культурою узбережжя Амальфі пропонує незабутні враження від подорожей. Якщо ви досліджуєте історичні вулиці Позітано, смакуєте смачні страви середземноморської кухні чи відпочиваєте на незайманих пляжах, кожна мить — це пам’ять, гідна листівки, яка чекає, щоб її зафіксувати."
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "public"
                                                            },
                                                            "posterUrl": {
                                                                "type": "string",
                                                                "example": "https://images-cdn.ubuy.co.in/633ff1157e3fbc25557517c8-one-piece-poster-japanese-anime-posters.jpg"
                                                            },
                                                            "bannerUrl": {
                                                                "type": "string",
                                                                "example": "https://i.pinimg.com/736x/00/a7/81/00a781cc93f26bc0b753e18b240673e2.jpg"
                                                            },
                                                            "likes": {
                                                                "type": "integer",
                                                                "example": "100"
                                                            },
                                                            "views": {
                                                                "type": "integer",
                                                                "example": "2266"
                                                            },
                                                            "wordsCount": {
                                                                "type": "integer",
                                                                "example": "41"
                                                            },
                                                            "wordsLearned": {
                                                                "type": "integer",
                                                                "example": "25"
                                                            },
                                                            "isStarted": {
                                                                "type": "boolean",
                                                                "example": "0"
                                                            },
                                                            "isLiked": {
                                                                "type": "boolean",
                                                                "example": "0"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Update user data",
                "operationId": "bf143d2b06c86da1bedae189ba91d052",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Maksym"
                                            },
                                            "surname": {
                                                "type": "string",
                                                "example": "Prokopenko"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Request successful"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "MaksymProkopenko@gmail.com"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Maksym"
                                                        },
                                                        "surname": {
                                                            "type": "string",
                                                            "example": "Prokopenko"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}